Adds splash screen containing the app's logo
In the future we might want to add the name of the app as well. It seems that there is no easy way to do this. Either we create an activity and use a layout, but this increases start-up times. An alternative would be to apply the app name as a path to the xml. In this case we would need to change the path at runtime depending on the user language